backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.54.0
python-multipart>=0.0.12
sqlalchemy>=2.0.36
pillow>=11.0.0
aiofiles>=24.1.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
